Rename mvp_example package to blueprints.mvp
