+++ mvp_example/src/main/java/com/distillery/android/blueprints/mvp/feature/todo/ToDoView.kt
@@ -4,9 +4,9 @@ import android.view.View
 import android.widget.EditText
 import android.widget.Toast
 import androidx.appcompat.app.AlertDialog
+import com.distillery.android.blueprints.mvp.R
 import com.distillery.android.blueprints.mvp.architecture.BaseView
 import com.distillery.android.domain.models.ToDoModel
-import com.distillery.android.mvp_example.R
 import com.distillery.android.ui.databinding.FragmentTodoBinding
 
 abstract class ToDoView(private val binding: FragmentTodoBinding) : BaseView {
